# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "API системы опросов для преддипломной практики",
        "contact": {
            "email": "vadim.starostin.07@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Локальный сервер"
        }
    ],
    "paths": {
...
        "/api/questions/{questionId}/options": {
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Добавить вариант ответа к вопросу",
                "operationId": "c17faa57dc224f8f3fb43831cffecf76",
                "parameters": [
                    {
                        "name": "questionId",
                        "in": "path",
                        "description": "ID вопроса (не текстового)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "option_text"
                                ],
                                "properties": {
                                    "option_text": {
                                        "type": "string",
                                        "example": "Stage 1 (300 л.с.)"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Вариант успешно добавлен"
                    },
                    "403": {
                        "description": "Доступ запрещен или опрос опубликован"
                    },
                    "422": {
                        "description": "Нельзя добавить вариант к текстовому вопросу"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/surveys": {
            "get": {
                "tags": [
                    "Surveys"
                ],
                "summary": "Получить список всех опросов",
                "operationId": "7e96246f34de1fcd663f2d95a5079073",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фильтр по статусу (draft, published, closed)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "my",
                        "in": "query",
                        "description": "Показать только мои опросы (1)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный список опросов"
                    }
                }
            },
            "post": {
                "tags": [
                    "Surveys"
                ],
                "summary": "Создать новый опрос (Только для админов)",
                "operationId": "a2cd6709081ad25c80c95f4cd8e97e2e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Тюнинг выхлопной системы"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Опрос о предпочтениях в звуке выхлопа"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Опрос успешно создан"
                    },
                    "403": {
                        "description": "Недостаточно прав"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/surveys/{id}": {
            "get": {
                "tags": [
                    "Surveys"
                ],
                "summary": "Детальная информация об опросе",
                "operationId": "be961dc9b72fd27b646853e31bdac701",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные опроса с вопросами"
                    }
                }
            }
        },
        "/api/surveys/{id}/status": {
            "patch": {
                "tags": [
                    "Surveys"
                ],
                "summary": "Смена статуса опроса",
                "operationId": "10f662df1396ade0cd460333fb18ca51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "published",
                                            "closed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Статус изменен"
                    },
                    "422": {
                        "description": "Нарушение жизненного цикла"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/surveys/{id}/passing": {
            "get": {
                "tags": [
                    "Surveys"
                ],
                "summary": "Получить опрос для прохождения (только опубликованные)",
                "operationId": "38daa5eef6320aa7149d6d4ce5f97227",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные для прохождения"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}